This reverts commit fd3486c

+++ src/BezahlAutomatIF.java
@@ -2,5 +2,5 @@ public interface BezahlAutomatIF {
 
     boolean bezahlen(ParkticketIF parkTicket);
 
-    float getPreis(ParkticketIF parkTicket);
+    Double getPreis(ParkticketIF parkTicket);
 }
